refactor logins
